Finish compiler unit tests
Everything is tested except the part that generates an Observable from chokidar events. I didn't know how to do that.

Lots of rearrangement - extracted several methods and functions for testing. Now the compile method is just a subscribe on an Observable that's set up in the constructor.
Change glob implementation to glob-promise. This lets us use much less code for creating the Observable.